Keep track of skip counts
This is done indirectly by keeping track of the number of games that
each user has been added to. Then to identify the number of skips, one
subtracts the total number of games the user has been added to by the
numer of turns that exist in the Turns table.

As a turn is deleted when a user is skipped, the difference will equal
the number of times the user has been skipped.

WIP #25