Create new history summary endpoint
This new endpoint returns a view of the entire user history. The
information that is given is significantly smaller than the full game
view and only includes what is needed to render a placeholder view of a
game in the history page.

Each individual game can be loaded asynchronously after the summaries
are available.