Attempt to improve login UX
If a user has not been logged in for a long time the app may go to the
inbox, load for a while, and then bump the user back to the login
screen.

Instead, try a lightweight request to ensure that the credentials are
still valid before attempting multiple slower queries.