Add an everything API endpoint
Looking at profiling data it looks like turn reaping is the really slow
part server side and that all other queries are pretty fast.

However, when running on Chrome and logging in, sometimes one of the
initial data fetch requests will hang for up to 7 seconds. This does not
appear to be reflected in any server side latency though, which suggests
that combining all of these concurrent requests together may fix this
client side latency.